requirements.txt
MetaTrader5
pandas
numpy
scikit-learn
joblib
psycopg2-binary
SQLAlchemy
pyyaml
python-dotenv
pytz
